fix: make it more clear that the user needs to set fbc_opt_in
Try to print out a more clear and instructive message for the user in
the event that the fbc_opt_in flag is not set to true.

Also, move these echo statements to the end of the task, so that they
are the obvious last line(s) of the taskrun log.

Signed-off-by: Ralph Bean <rbean@redhat.com>