fix: properly expand create-pyxis-image oras args
We were setting the ORAS_ARGS based on variables in all lowercase, but
the variables are actually capitalized. The reason why we never hit this
until now is because the if statement that sets these variables was
never triggered until a typo was fixed in the if condition with the
linting fix yesterday.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>