fix: composing oras args in create-pyxis-image
This line was just modified earlier today to fix the os and arch
variables. But it is still wrong - you want the name of
argument and the actual value separate, otherwise the whole
thing will be used as one parameter.

Since it's still the same line and the changelog entry
still applies to this, I don't think I need to bump
the version again.

Signed-off-by: Martin Malina <mmalina@redhat.com>