feat(RELEASE-1211): add push workflow so internal mirrors development
This commit adds a github workflow to push to the internal branch
whenever a push is made to the development branch. Thus, the internal
branch will mirror the development branch so internal can be used for
the app-sre cluster. It also modifies the promote_branch script to no
longer push to the internal branch and explains the PR process for the
internal directory in CONTRIBUTING.md.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>