hack: workaround for bootc
Do not merge. This is a temporary pipeline to for rpm-ostree images.

The image update here pulls in this change:
https://github.com/konflux-ci/release-service-utils/pull/311

It makes it so that in the event that the layer data is absent, a `null`
value won't be sent to pyxis which would be rejected by pyxis' input
validation.