feat(RELEASE-1336): skip image layers in Pyxis with skipLayers
A new option is added to data file to be able to skip layer
details when create Pyxis image.

This was previously created as a one-off workaround to unblock
bootc releases here:
https://github.com/konflux-ci/release-service-catalog/pull/684

Now it's an option that can be enabled as needed.

For full explanation, see the jira and also the comment
in the task script.

Signed-off-by: Martin Malina <mmalina@redhat.com>