feat(SPSTRAT-482): add pre-push for marketplacesvm
This commit allows the `marketplacesvm-push-disk-images` task to set the
parameter `nochannel` on the wrapper to use the `pre-push` mode.

The `pre-push` mode will cause the VM images to be uploaded to the
respective clouds but it won't fully publish on them. This is ideal to
perform the "staging" tests before the final push.

It also fixes a naming issue with the task
`collect-marketplacesvm-secret`.

Refers to SPSTRAT-482

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>