feat(RELEASE-1347): add resources to update manager yaml files
This commit adds a tenant task to fetch the image ref containing a git
sha tag from the Release artifacts as well as a tenant task to update a
file in git with the value of the image ref.

It also adds a tenant pipeline that calls these two tasks.

The point of this pipeline is to use it as a final pipeline. It will
enable us to have the internal-services manager files updated once the
internal-services managed pipelineRun succeeds. Today, this is a manual
step.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>