refactor: specify branches instead of target branch in promotion script
Previously what was being promoted was potentially confusing because
the description said "which branch to promote", but it was actually
the target branch, not the source branch. And we promote the source
branch to the target branch.

I propose to specify both branches at once, i.e.
development-to-staging or staging-to-production to make it clearer.

Signed-off-by: Martin Malina <mmalina@redhat.com>