refactor(RELEASE-1374): split out pulling of artifacts as a step
The push-artifacts-to-cdn-task originally did all in one step -
pulling of the artifacts via oras and then pushing them to cdn.

This change splits the pulling out to a separate step. This is
in preparation of adding another step in between for signing.

An emptyDir is used for sharind data among the steps.

Signed-off-by: Martin Malina <mmalina@redhat.com>