feat(RELEASE-1400): add references in populate-release-notes task
This commit renames the populate-release-notes-images task to
populate-release-notes. The new version of the task adds references if
the type is RHSA. One is a static reference and there is also a
reference added per CVE listed in releaseNotes.

Because references will no longer need to be passed for RHBA or RHEA as
part of this work, this commit also updates the schema to not require
references when using releaseNotes.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>