feat(RELEASE-1403): rework embargo-check task for issues
This commit reworks the managed embargo-check task for how it handles
issues in issues.redhat.com. First, it adds authentication for searching
for these issues. If the issue is found and is not of type
Vulnerability, it will pass.

If the issue is of type Vulnerability, the task will now ensure that the
component from the issue matches a component in the
releaseNotes.content.images section and that the CVE from the issue is
present in said component section.

Signed-off-by: Johnny Bieren <jbieren@redhat.com>